requirements.txt
streamlit
openai
huggingface_hub
diffusers
torch
python-dotenv